feat: description ui
