+++ src/app/(routes)/collections/create/page.tsx
@@ -1,19 +1,9 @@
-import SelectQuizFromCollection from '@/features/collection/components/select-quiz-from-collection'
-import Text from '@/shared/components/ui/text'
+import CreateCollectionForm from '@/features/collection/components/create-collection-form'
 
 const CreateCollectionPage = () => {
   return (
     <div className="px-[20px]">
-      <div className="mt-[27px]">
-        <Text typography="title3" className="text-text-primary">
-          컬렉션으로 만들 문제를 선택해주세요
-        </Text>
-        <Text typography="text1-medium" className="mt-[8px] text-text-sub">
-          5문제 이상 선택해주세요.
-        </Text>
-      </div>
-
-      <SelectQuizFromCollection />
+      <CreateCollectionForm />
     </div>
   )
 }
